Exercise 2.16 (Arithmetic)
Program Tasks: Write a program that asks the user to enter two numbers,
obtain the two numbers from the user and print the sum, product,
difference, and quotient of the two numbers.

Programmer Outcome: Apon writing this program i ran into an error with
using modulus to return a remainder while using the quotient formula.
More input could be made to see which number is bigger to then complete
the calculations so more precise information can be returned to the
user, as well as less code could be written to better clean up the
code. more comments can be placed in the program to direct the user on
how to achieve these same calculations as well as giving two different
sets of answers.