Chapter 3 Examples
read through the chapter twice before writing code to see it function.
Began on exercise 3.11 to modify grade book with instructor. ran into
some problems going to sit on it for the night and try to re work out
the problem. Started to understand how to use Xcode to create header
files separate from the main file. Had to spend some time learning how
to use the IDE.