personal project to practice skills.
trying to work on the skills done in class, not to much luck on it,
however a lot of code placed in the main cpp file that can be used for
the long run. looking to manage and reduce some of the clutter into
smaller more manageable code to put the program to full use.